Make KeyWeakMap a template class
