+++ filenames.gypi
@@ -340,6 +340,7 @@
       'atom/common/google_api_key.h',
       'atom/common/id_weak_map.cc',
       'atom/common/id_weak_map.h',
+      'atom/common/key_weak_map.h',
       'atom/common/keyboard_util.cc',
       'atom/common/keyboard_util.h',
       'atom/common/mouse_util.cc',
